# Improve SerializationVisitor performance