--- a/src/NCalc.Core/NCalc.Core.csproj
+++ b/src/NCalc.Core/NCalc.Core.csproj
@@ -37,6 +37,7 @@
         <PackageReference Include="Microsoft.Bcl.AsyncInterfaces" Version="8.0.0" />
         <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="8.0.1" />
         <PackageReference Include="Microsoft.Extensions.Logging.TraceSource" Version="8.0.1" />
+        <PackageReference Include="Microsoft.Extensions.ObjectPool" Version="8.0.10" />
         <PackageReference Include="Parlot" Version="1.0.2" />
     </ItemGroup>
 </Project>