--- a/src/NCalc.Core/Visitors/SerializationVisitor.cs
+++ b/src/NCalc.Core/Visitors/SerializationVisitor.cs
@@ -15,148 +15,130 @@ public class SerializationVisitor : ILogicalExpressionVisitor<string>
 
     public string Visit(TernaryExpression expression)
     {
-        var result = new StringBuilder();
-        result.Append(EncapsulateNoValue(expression.LeftExpression));
-        result.Append("? ");
-        result.Append(EncapsulateNoValue(expression.MiddleExpression));
-        result.Append(": ");
-        result.Append(EncapsulateNoValue(expression.RightExpression));
-        return result.ToString();
+        string result = EncapsulateNoValue(expression.LeftExpression) + "? ";
+        result += EncapsulateNoValue(expression.MiddleExpression) + ": ";
+        result += EncapsulateNoValue(expression.RightExpression);
+
+        return result;
     }
 
     public string Visit(BinaryExpression expression)
     {
-        var result = new StringBuilder();
-        result.Append(EncapsulateNoValue(expression.LeftExpression));
+        string result = EncapsulateNoValue(expression.LeftExpression);
 
         switch (expression.Type)
         {
             case BinaryExpressionType.And:
-                result.Append("and ");
+                result += "and ";
                 break;
             case BinaryExpressionType.Or:
-                result.Append("or ");
+                result += "or ";
                 break;
             case BinaryExpressionType.Div:
-                result.Append("/ ");
+                result += "/ ";
                 break;
             case BinaryExpressionType.Equal:
-                result.Append("= ");
+                result += "= ";
                 break;
             case BinaryExpressionType.Greater:
-                result.Append("> ");
+                result += "> ";
                 break;
             case BinaryExpressionType.GreaterOrEqual:
-                result.Append(">= ");
+                result += ">= ";
                 break;
             case BinaryExpressionType.Lesser:
-                result.Append("< ");
+                result += "< ";
                 break;
             case BinaryExpressionType.LesserOrEqual:
-                result.Append("<= ");
+                result += "<= ";
                 break;
             case BinaryExpressionType.Minus:
-                result.Append("- ");
+                result += "- ";
                 break;
             case BinaryExpressionType.Modulo:
-                result.Append("% ");
+                result += "% ";
                 break;
             case BinaryExpressionType.NotEqual:
-                result.Append("!= ");
+                result += "!= ";
                 break;
             case BinaryExpressionType.Plus:
-                result.Append("+ ");
+                result += "+ ";
                 break;
             case BinaryExpressionType.Times:
-                result.Append("* ");
+                result += "* ";
                 break;
             case BinaryExpressionType.BitwiseAnd:
-                result.Append("& ");
+                result += "& ";
                 break;
             case BinaryExpressionType.BitwiseOr:
-                result.Append("| ");
+                result += "| ";
                 break;
             case BinaryExpressionType.BitwiseXOr:
-                result.Append("^ ");
+                result += "^ ";
                 break;
             case BinaryExpressionType.LeftShift:
-                result.Append("<< ");
+                result += "<< ";
                 break;
             case BinaryExpressionType.RightShift:
-                result.Append(">> ");
+                result += ">> ";
                 break;
             case BinaryExpressionType.Exponentiation:
-                result.Append("** ");
+                result += "** ";
                 break;
             case BinaryExpressionType.In:
-                result.Append("in ");
+                result += "in ";
                 break;
             case BinaryExpressionType.NotIn:
-                result.Append("not in ");
+                result += "not in ";
                 break;
             case BinaryExpressionType.Like:
-                result.Append("like ");
+                result += "like ";
                 break;
             case BinaryExpressionType.NotLike:
-                result.Append("not like ");
+                result += "not like ";
                 break;
             case BinaryExpressionType.Unknown:
-                result.Append("unknown ");
+                result += "unknown ";
                 break;
             default:
                 throw new ArgumentOutOfRangeException();
         }
 
-        result.Append(EncapsulateNoValue(expression.RightExpression));
-        return result.ToString();
+        result += EncapsulateNoValue(expression.RightExpression);
+        return result;
     }
 
     public string Visit(UnaryExpression expression)
     {
-        var result = new StringBuilder();
+        string result = "";
 
         switch (expression.Type)
         {
             case UnaryExpressionType.Not:
-                result.Append('!');
+                result = "!";
                 break;
             case UnaryExpressionType.Negate:
-                result.Append('-');
+                result = "-";
                 break;
             case UnaryExpressionType.BitwiseNot:
-                result.Append('~');
+                result = "~";
                 break;
         }
 
-        result.Append(EncapsulateNoValue(expression.Expression));
-        return result.ToString();
+        result += EncapsulateNoValue(expression.Expression);
+        return result;
     }
 
     public string Visit(ValueExpression expression)
     {
-        var result = new StringBuilder();
-
-        switch (expression.Type)
+        return expression.Type switch
         {
-            case ValueType.Boolean:
-                result.Append(expression.Value).Append(' ');
-                break;
-            case ValueType.DateTime or ValueType.TimeSpan:
-                result.Append('#').Append(expression.Value).Append('#').Append(' ');
-                break;
-            case ValueType.Float:
-                result.Append(decimal.Parse(expression.Value?.ToString() ?? string.Empty).ToString(_numberFormatInfo))
-                    .Append(' ');
-                break;
-            case ValueType.Integer:
-                result.Append(expression.Value).Append(' ');
-                break;
-            case ValueType.String or ValueType.Char:
-                result.Append('\'').Append(expression.Value).Append('\'').Append(' ');
-                break;
-        }
-
-        return result.ToString();
+            ValueType.Boolean or ValueType.Integer => $"{expression.Value} ",
+            ValueType.DateTime or ValueType.TimeSpan => $"#{expression.Value}# ",
+            ValueType.Float => $"{decimal.Parse(expression.Value?.ToString() ?? string.Empty).ToString(_numberFormatInfo)} ",
+            ValueType.String or ValueType.Char => $"'{expression.Value}' ",
+            _ => "",
+        };
     }
 
     public string Visit(Function function)
@@ -183,9 +165,7 @@ public class SerializationVisitor : ILogicalExpressionVisitor<string>
 
     public string Visit(Identifier identifier)
     {
-        var result = new StringBuilder();
-        result.Append('[').Append(identifier.Name).Append("] ");
-        return result.ToString();
+        return $"[{identifier.Name}] ";
     }
 
     public string Visit(LogicalExpressionList list)
@@ -211,15 +191,7 @@ public class SerializationVisitor : ILogicalExpressionVisitor<string>
             return valueExpression.Accept(this);
         }
 
-        var result = new StringBuilder();
-        result.Append('(');
-        result.Append(expression.Accept(this));
-
-        // trim spaces before adding a closing paren
-        while (result[^1] == ' ')
-            result.Remove(result.Length - 1, 1);
-
-        result.Append(") ");
-        return result.ToString();
+        string result = expression.Accept(this);
+        return $"({result.TrimEnd(' ')}) ";
     }
 }