This PR improves `SerializationVisitor` performance. 

I did some benchmarking and `StringBuilder` usage in some case is not the best solution, here some benchmark results:

```
[MemoryDiagnoser]
public class TestComparision
{
	public const string identifier = "identifier";

	[Benchmark(Baseline = true)]
	public string TestStringBuilder()
	{
		var result = new StringBuilder();
		result.Append('(');
		result.Append(identifier);

		// trim spaces before adding a closing paren
		while (result[^1] == ' ')
			result.Remove(result.Length - 1, 1);

		result.Append(") ");
		return result.ToString();
	}

	[Benchmark]
	public string TestStringConcat()
	{
		return "(" + identifier.TrimEnd(' ') + ") ";
	}

	[Benchmark]
	public string TestStringInterpolation()
	{
		return $"({identifier.TrimEnd(' ')}) ";
	}
}
```

BenchmarkDotNet v0.14.0, Windows 10 (10.0.19045.5011/22H2/2022Update)
Intel Core i3-6100U CPU 2.30GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
.NET SDK 8.0.206
  [Host]     : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2


| Method                  | Mean     | Error    | StdDev   | Median   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|------------------------ |---------:|---------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| TestStringBuilder       | 50.02 ns | 1.114 ns | 1.598 ns | 49.23 ns |  1.00 |    0.04 | 0.0969 |     152 B |        1.00 |
| TestStringConcat        | 26.80 ns | 0.228 ns | 0.190 ns | 26.74 ns |  0.54 |    0.02 | 0.0306 |      48 B |        0.32 |
| TestStringInterpolation | 27.64 ns | 0.625 ns | 1.647 ns | 26.73 ns |  0.55 |    0.04 | 0.0306 |      48 B |        0.32 |

  
```
[MemoryDiagnoser]
public class TestComparision
{
	public const string identifier = "identifier";

	[Benchmark(Baseline = true)]
	public string TestStringBuilderLoop()
	{
		var result = new StringBuilder();

		for (int i = 0; i < 5; i++)
			result.Append(identifier);

		return result.ToString();
	}

	[Benchmark]
	public string TestStringConcatLoop()
	{
		string str = "";
		for (int i = 0; i < 5; i++)
			str += identifier;

		return str;
	}
}
```
  BenchmarkDotNet v0.14.0, Windows 10 (10.0.19045.5011/22H2/2022Update)
Intel Core i3-6100U CPU 2.30GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
.NET SDK 8.0.206
  [Host]     : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2


| Method                | Mean      | Error    | StdDev   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|---------------------- |----------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| TestStringBuilderLoop | 195.10 ns | 2.659 ns | 2.487 ns |  1.00 |    0.02 | 0.3009 |     472 B |        1.00 |
| TestStringConcatLoop  |  97.86 ns | 1.128 ns | 0.881 ns |  0.50 |    0.01 | 0.2447 |     384 B |        0.81 |