@gumbarros 2024-10-30
Nice job!

Before merging can you check performance using `StringBuilderPool`? You need to install `Microsoft.Extensions.ObjectPool`, the reason of the slow down is probably because multiple `StringBuilder` instances are created.

Example of usage:

https://github.com/JJConsulting/JJMasterData/blob/c82ed59d2cf66c938e941c093022a01b21229180/src/Core/UI/Html/HtmlBuilder.cs#L158