@Bykiev 2024-10-31
Thank you, I've never heard about `StringBuilderPool` before and it was interesting to touch it!
I did some benchmarks and as expected the result is much better in cases when there are a lot of repeating strings. `SerializationVisitor` looks like to be a good candidate for it. Should I revert all my changes and just replace `StringBuilder` with `StringBuilderPool`?

Here is results of benchmarks:

```
    [MemoryDiagnoser]
    public class TestComparision
    {
        public const string identifier = "identifier";

        private static readonly ObjectPool<StringBuilder> StringBuilderPool;

        static TestComparision()
        {
            StringBuilderPool = new DefaultObjectPoolProvider().CreateStringBuilderPool();
        }

        [Benchmark(Baseline = true)]
        public string TestStringBuilder()
        {
            var result = new StringBuilder();
            result.Append('(');
            result.Append(identifier);

            // trim spaces before adding a closing paren
            while (result[^1] == ' ')
                result.Remove(result.Length - 1, 1);

            result.Append(") ");
            return result.ToString();
        }

        [Benchmark]
        public string TestStringBuilderPool()
        {
            var result = StringBuilderPool.Get();
            result.Append('(');
            result.Append(identifier);

            // trim spaces before adding a closing paren
            while (result[^1] == ' ')
                result.Remove(result.Length - 1, 1);

            result.Append(") ");
            StringBuilderPool.Return(result);

            return result.ToString();
        }

        [Benchmark]
        public string TestStringConcat()
        {
            return "(" + identifier.TrimEnd(' ') + ") ";
        }
    }
```

// * Summary *

BenchmarkDotNet v0.14.0, Windows 10 (10.0.19045.5011/22H2/2022Update)
Intel Core i3-6100U CPU 2.30GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
.NET SDK 8.0.206
  [Host]     : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2


| Method                | Mean     | Error    | StdDev   | Median   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|---------------------- |---------:|---------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| TestStringBuilder     | 91.18 ns | 3.327 ns | 9.811 ns | 94.26 ns |  1.01 |    0.16 | 0.0969 |     152 B |        1.00 |
| TestStringBuilderPool | 75.22 ns | 1.540 ns | 1.711 ns | 75.27 ns |  0.84 |    0.10 |      - |         - |        0.00 |
| TestStringConcat      | 55.13 ns | 0.479 ns | 0.400 ns | 55.24 ns |  0.61 |    0.07 | 0.0305 |      48 B |        0.32 |

// * Hints *
Outliers
  TestComparision.TestStringBuilderPool: Default -> 1 outlier  was  removed, 3 outliers were detected (79.21 ns, 79.89 ns, 86.76 ns)
  TestComparision.TestStringConcat: Default      -> 2 outliers were removed (63.23 ns, 64.02 ns)
  
  
  ```

    [MemoryDiagnoser]
    public class TestComparision
    {
        public const string identifier = "identifier";

        private static readonly ObjectPool<StringBuilder> StringBuilderPool;

        static TestComparision()
        {
            StringBuilderPool = new DefaultObjectPoolProvider().CreateStringBuilderPool();
        }

        [Benchmark(Baseline = true)]
        public string TestStringBuilderLoop()
        {
            var result = new StringBuilder();

            for (int i = 0; i < 5; i++)
                result.Append(identifier);

            return result.ToString();
        }

        [Benchmark]
        public string TestStringBuilderPoolLoop()
        {
            var result = StringBuilderPool.Get();

            for (int i = 0; i < 5; i++)
                result.Append(identifier);

            StringBuilderPool.Return(result);

            return result.ToString();
        }

        [Benchmark]
        public string TestStringConcatLoop()
        {
            string str = "";
            for (int i = 0; i < 5; i++)
                str += identifier;

            return str;
        }
    }
```
  
  // * Summary *

BenchmarkDotNet v0.14.0, Windows 10 (10.0.19045.5011/22H2/2022Update)
Intel Core i3-6100U CPU 2.30GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
.NET SDK 8.0.206
  [Host]     : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2


| Method                    | Mean      | Error    | StdDev   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|-------------------------- |----------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
| TestStringBuilderLoop     | 227.25 ns | 4.697 ns | 8.823 ns |  1.00 |    0.05 | 0.3009 |     472 B |        1.00 |
| TestStringBuilderPoolLoop |  82.78 ns | 1.752 ns | 2.456 ns |  0.36 |    0.02 |      - |         - |        0.00 |
| TestStringConcatLoop      | 119.03 ns | 2.415 ns | 4.709 ns |  0.52 |    0.03 | 0.2446 |     384 B |        0.81 |

```
[MemoryDiagnoser]
public class TestComparision
{
    private static readonly ObjectPool<StringBuilder> StringBuilderPool;

    static TestComparision()
    {
        StringBuilderPool = new DefaultObjectPoolProvider().CreateStringBuilderPool();
    }

    [Benchmark(Baseline = true)]
    public string TestStringBuilderLoop()
    {
        var result = new StringBuilder();

        for (int i = 0; i < 5; i++)
            result.Append(Random.Shared.NextDouble());

        return result.ToString();
    }

    [Benchmark]
    public string TestStringBuilderPoolLoop()
    {
        var result = StringBuilderPool.Get();

        for (int i = 0; i < 5; i++)
            result.Append(Random.Shared.NextDouble());

        StringBuilderPool.Return(result);

        return result.ToString();
    }

    [Benchmark]
    public string TestStringConcatLoop()
    {
        string str = "";
        for (int i = 0; i < 5; i++)
            str += Random.Shared.NextDouble();

        return str;
    }
}
```

BenchmarkDotNet v0.14.0, Windows 10 (10.0.19045.5011/22H2/2022Update)
Intel Core i3-6100U CPU 2.30GHz (Skylake), 1 CPU, 4 logical and 2 physical cores
.NET SDK 8.0.206
  [Host]     : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.35 (6.0.3524.45918), X64 RyuJIT AVX2


| Method                    | Mean     | Error     | StdDev    | Median   | Ratio | RatioSD | Gen0   | Allocated | Alloc Ratio |
|-------------------------- |---------:|----------:|----------:|---------:|------:|--------:|-------:|----------:|------------:|
| TestStringBuilderLoop     | 2.999 us | 0.0250 us | 0.0221 us | 2.991 us |  1.00 |    0.01 | 0.5989 |     942 B |        1.00 |
| TestStringBuilderPoolLoop | 1.722 us | 0.0346 us | 0.0751 us | 1.694 us |  0.57 |    0.03 |      - |         - |        0.00 |
| TestStringConcatLoop      | 1.890 us | 0.0338 us | 0.0282 us | 1.880 us |  0.63 |    0.01 | 0.5913 |     928 B |        0.99 |

// * Hints *
Outliers
  TestComparision.TestStringBuilderLoop: Default     -> 1 outlier  was  removed (3.22 us)
  TestComparision.TestStringBuilderPoolLoop: Default -> 7 outliers were removed (2.04 us..2.88 us)
  TestComparision.TestStringConcatLoop: Default      -> 2 outliers were removed (2.08 us, 2.17 us)